requirements.txt
Django==3.1
Pillow==7.1
djangorestframework==3.11
python-decouple==3.4
drf-yasg
psycopg2-binary==2.8.5
BeautifulSoup4
lxml
requests
